requirements.txt
flask
Jinja2
gunicorn
pylint
requests
pyyaml
python-dotenv
SQLAlchemy==1.4.39